refactor: removed MessageContext and moved logic to local state in ChatBoxList
