+++ client/src/components/ChatBox/ChatBox.tsx
@@ -1,33 +1,22 @@
-import { useContext, useEffect, useState, useRef } from 'react';
-import { MessageContext } from '../../context/MessageContext';
+import { useContext, useEffect, useRef } from 'react';
 import { ChatContext } from '../../context/ChatContext';
-import {
-  calculateLeft,
-  calculateTop,
-  getRandomColor,
-} from '../../utils';
 import ChatBoxInput from './ChatBoxInput';
 import ChatBoxHeader from './ChatBoxHeader';
-import { ChatContextType, MessageContextType } from '../../context/ContextTypes';
+import { ChatContextType } from '../../context/ContextTypes';
 import useWindowMove from '../../hooks/useWindowMove';
+import { Message } from '../types';
 
 type IChatBoxProps = {
+  messageList: Message[],
+  sendMessage: (roomName: string, message: string) => void;
   roomName: string,
   socketId: string,
   handleBackgroundColor: () => void;
 };
 
-export function ChatBox ({ roomName, socketId, handleBackgroundColor }: IChatBoxProps) {
-  const { messageList, sendMessage } = useContext(MessageContext) as MessageContextType;
+export function ChatBox ({ messageList, sendMessage, roomName, socketId, handleBackgroundColor }: IChatBoxProps) {
   const { leaveRoom } = useContext(ChatContext) as ChatContextType;
   const { position, handleMouseDown, handleMouseMove, handleMouseUp } = useWindowMove();
-
-  //Color states
-  // const [colorMap, setColorMap] = useState({});
-  // const [color, setColor] = useState(
-  //   '#' + ((Math.random() * 0xffffff) << 0).toString(16)
-  // );
-  // Define the color variable
 
   // MESSAGE FUNCTIONALITY
   const handleSendMessage = (message: string) => {
@@ -39,30 +28,6 @@ export function ChatBox ({ roomName, socketId, handleBackgroundColor }: IChatBox
     leaveRoom(roomName);
   };
 
-  // Colors of the usernames
-  // useEffect(() => {
-  //   setColorMap((prevColorMap) => {
-  //     return {
-  //       ...prevColorMap,
-  //       [socketId]: color,
-  //     };
-  //   });
-  // }, [socketId, color]);
-
-  // function getColor (socketId: string) {
-  //   if (!colorMap[socketId]) {
-  //     // Generate a random color for new users
-  //     setColorMap((prevColorMap) => {
-  //       return {
-  //         ...prevColorMap,
-  //         [socketId]: getRandomColor(),
-  //       };
-  //     });
-  //   }
-  //   return colorMap[socketId];
-  // }
-
-  // Auto scroll when there is a new message
   const messagesEndRef = useRef<HTMLDivElement>(null);
   const scrollToBottom = () => {
     messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
@@ -70,7 +35,6 @@ export function ChatBox ({ roomName, socketId, handleBackgroundColor }: IChatBox
   useEffect(() => {
     scrollToBottom();
   }, [messageList]);
-
 
   return (
     <>
@@ -109,4 +73,38 @@ export function ChatBox ({ roomName, socketId, handleBackgroundColor }: IChatBox
       </div>
     </>
   );
+
+  // LOGIC to assign different colors to usernames
+
+  //Color states
+  // const [colorMap, setColorMap] = useState({});
+  // const [color, setColor] = useState(
+  //   '#' + ((Math.random() * 0xffffff) << 0).toString(16)
+  // );
+  // Define the color variable
+
+  // Colors of the usernames
+  // useEffect(() => {
+  //   setColorMap((prevColorMap) => {
+  //     return {
+  //       ...prevColorMap,
+  //       [socketId]: color,
+  //     };
+  //   });
+  // }, [socketId, color]);
+
+  // function getColor (socketId: string) {
+  //   if (!colorMap[socketId]) {
+  //     // Generate a random color for new users
+  //     setColorMap((prevColorMap) => {
+  //       return {
+  //         ...prevColorMap,
+  //         [socketId]: getRandomColor(),
+  //       };
+  //     });
+  //   }
+  //   return colorMap[socketId];
+  // }
+
+  // Auto scroll when there is a new message
 }
