+++ client/src/components/types.ts
@@ -1,0 +1,6 @@
+export type Message = {
+  user: string;
+  roomName: string;
+  message: string;
+  time: string;
+};
