+++ client/src/context/ContextTypes.tsx
@@ -32,15 +32,3 @@ export type ChatContextType = {
   isSelectorClosed: boolean,
   isSelectorVisible: boolean,
 };
-
-export type Message = {
-  user: string;
-  roomName: string;
-  message: string;
-  time: string;
-};
-
-export type MessageContextType = {
-  messageList: Message[];
-  sendMessage: (roomName: string, message: string) => void;
-};
